- Improved the -W (--fixed-width) option to use `Unicode::GCString` if available (in addition to `Text::CharWidth`).
A fifth of the terminal width is now allocated for the channel names, so very long channel names will no longer distort the results.
Additionally, the channel names are now formated in the same way as the titles, which fixes some issues with channel names that contain grapheme clusters of width=1.