- Added support for local playlists (CLI). (many thanks to @rgasc)
Playlists are lists of video IDs (one video ID per line) and are stored in `~/.config/youtube-viewer/playlists/`

Example:

	youtube-viewer -lp=playlistName

The `playlistName` can be a substring of a playlist filename located in the playlist directory.

Playlists can be easily created with `youtube-dl`:

	youtube-dl --get-id PLAYLIST_URL -i >> ~/.config/youtube-viewer/playlists/my_playlist

Implements and closes https://github.com/trizen/youtube-viewer/issues/345.