- Added the `-a` alias for `--audio`. (closes https://github.com/trizen/youtube-viewer/issues/380)
However, for playing only audio, the `-n` flag should be preferred instead (which implies `--audio`).

The difference is that `-n` will not fail if there is no separate audio track available. Instead, it will simply disables the video playback (by passing `--no-video` to `mpv`).

When there is an audio track availalble, `-a` and `-n` are equivalent.