- Added the `subscription_videos_per_channel` config-option.
It specifies the number of video results to fetch from each channel when displaying local subscription videos.

Valid values are between 1 and 50. Previously, the default value was 10. Now, the default value is 20.

When a channel publishes more than this many new videos (since last check), only the most N recent ones are fetched.